Refactor of Dataset into seperate class.  Everything now loads.
